moved namespaces and added auto mapper
