+++ ChoreBuddies.Backend/Program.cs
@@ -1,88 +1,90 @@
-using ChoreBuddies.Backend.Chore;
-using ChoreBuddies.Backend.Tasks;
+using AutoMapper;
+using ChoreBuddies.Backend.features.Chores;
 using ChoreBuddies.Database;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
 using System;
 
 namespace ChoreBuddies.Backend;
 
 public class Program
 {
-    public async static Task Main(string[] args)
-    {
-        var builder = WebApplication.CreateBuilder(args);
-        builder.AddServiceDefaults();
+	public async static Task Main(string[] args)
+	{
+		var builder = WebApplication.CreateBuilder(args);
+		builder.AddServiceDefaults();
 
-        builder.Services.AddCors(options =>
-        {
-            options.AddDefaultPolicy(builder =>
-            {
-                builder.AllowAnyOrigin()
-                .AllowAnyMethod()
-                .AllowAnyHeader();
-            });
-        });
+		builder.Services.AddCors(options =>
+		{
+			options.AddDefaultPolicy(builder =>
+			{
+				builder.AllowAnyOrigin()
+				.AllowAnyMethod()
+				.AllowAnyHeader();
+			});
+		});
 
-        builder.Services.AddEndpointsApiExplorer();
+		builder.Services.AddEndpointsApiExplorer();
 
-        builder.Services.AddDbContext<ChoreBuddiesDbContext>(opt =>
-        {
-            opt.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection"));
-        });
+		builder.Services.AddDbContext<ChoreBuddiesDbContext>(opt =>
+		{
+			opt.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection"));
+		});
 
-        //builder.Services.AddMvc();
-        builder.Services.AddSwaggerGen();
+		//builder.Services.AddMvc();
+		builder.Services.AddSwaggerGen();
 
-        builder.Services.AddControllers();
+		builder.Services.AddControllers();
 
-        // Add services to the container.
-        builder.Services.AddRazorPages();
-        builder.Services.AddScoped<IChoresService, ChoresService>();
+		// Add services to the container.
+		builder.Services.AddRazorPages();
+		builder.Services.AddScoped<IChoresService, ChoresService>();
+		builder.Services.AddAutoMapper(cfg => cfg.AddProfile<MappingProfile>());
 
-        var app = builder.Build();
+		var app = builder.Build();
 
-        // Apply migrations to database
-        await using (var serviceScope = app.Services.CreateAsyncScope())
-        await using (var dbContext = serviceScope.ServiceProvider.GetRequiredService<ChoreBuddiesDbContext>())
-        {
-            var pendingMigrations = await dbContext.Database.GetPendingMigrationsAsync();
+		// Apply migrations to database
+		await using (var serviceScope = app.Services.CreateAsyncScope())
+		await using (var dbContext = serviceScope.ServiceProvider.GetRequiredService<ChoreBuddiesDbContext>())
+		{
+			var pendingMigrations = await dbContext.Database.GetPendingMigrationsAsync();
 
-            if (pendingMigrations.Any())
-            {
-                await dbContext.Database.MigrateAsync();
-            }
-        }
+			if (pendingMigrations.Any())
+			{
+				await dbContext.Database.MigrateAsync();
+			}
+		}
 
-        app.MapDefaultEndpoints();
+		app.MapDefaultEndpoints();
 
-        app.MapControllers();
+		app.MapControllers();
 
-        // Configure the HTTP request pipeline.
-        if (!app.Environment.IsDevelopment())
-        {
-            app.UseExceptionHandler("/Error");
-            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-            app.UseHsts();
-        }
+		// Configure the HTTP request pipeline.
+		if (!app.Environment.IsDevelopment())
+		{
+			app.UseExceptionHandler("/Error");
+			// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+			app.UseHsts();
+		}
 
-        app.UseHttpsRedirection();
-        app.UseStaticFiles();
+		app.UseHttpsRedirection();
+		app.UseStaticFiles();
 
-        app.UseRouting();
+		app.UseRouting();
 
-        app.UseCors(policy => policy
-            .AllowAnyOrigin()
-            .AllowAnyMethod()
-            .AllowAnyHeader()
-        );
+		app.UseCors(policy => policy
+			.AllowAnyOrigin()
+			.AllowAnyMethod()
+			.AllowAnyHeader()
+		);
 
-        app.UseAuthorization();
+		app.UseAuthorization();
 
-        app.MapRazorPages();
+		app.MapRazorPages();
 
-        app.UseSwagger();
-        app.UseSwaggerUI();
+		app.UseSwagger();
+		app.UseSwaggerUI();
 
-        app.Run();
-    }
+		app.Run();
+	}
 }
