+++ ChoreBuddies.Backend/features/Chores/ChoreDto.cs
@@ -1,0 +1,15 @@
+﻿using ChoreBuddies.Backend.features.Chores;
+
+namespace ChoreBuddies.Backend.Chores
+{
+	public record ChoreDto(
+	string Id,
+	string Name,
+	string Description,
+	string? AssignedTo,
+	DateTime DueDate,
+	Status Status,
+	string Room,
+	int RewardPointsCount = 0
+);
+}
