+++ ChoreBuddies.Backend/features/Chores/Status.cs
@@ -1,4 +1,4 @@
-﻿namespace ChoreBuddies.Backend.Tasks
+﻿namespace ChoreBuddies.Backend.features.Chores
 {
 	public enum Status
 	{
